test(pts `ActivitiesController`): add tests to cover frequency dto integration & fix breaking changes due to controller changes
